Add: placement prop to InvoicesPDF component
