+++ src/helpers/index.ts
@@ -4,3 +4,4 @@ export * from './campaignDetails'
 export * from './timeout'
 export * from './balances'
 export * from './campaignPeriodParser'
+export * from './campaignAnalytics'
